requirements.txt
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
Brotli==1.1.0
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cookiecutter==2.6.0
cssselect2==0.8.0
dj-database-url==3.0.0
Django==5.2.3
django-allauth==65.9.0
django-tailwind==4.0.1
djangorestframework==3.16.0
fonttools==4.58.2
gunicorn==23.0.0
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.9
pycparser==2.22
pydyf==0.11.0
Pygments==2.19.2
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
tinycss2==1.4.0
tinyhtml5==2.0.0
types-python-dateutil==2.9.0.20250516
typing_extensions==4.14.0
urllib3==2.5.0
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
